+++ src/utils/enums.util.ts
@@ -1,4 +1,0 @@
-export enum Commands {
-    generate,
-    create
-}