+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "Serverpod" extension will be documented in this file.
 
+## [Pre-Release - 0.0.5]
+
+- Updated registering commands.
+- Added Unofficial Serverpod tag to README.md
+
 ## [Pre-Release - 0.0.4]
 
 - Added a feature to generate and watch client side code.